Move soe logic in soe package
